requirements.txt
requests==2.31.0
python-dotenv==1.0.0
ollama==0.3.3
schedule==1.2.0
colorama==0.4.6
